aegis/agents/integration.py
"""
Integration & FinOps Agent (Agent2).

Responsibilities:
- Integration marketplace recommendations
- API gateway configuration assistance
- Code generation for common integrations
"""

import logging
from typing import Optional

from aegis.agents.base import AgentMessage, AgentResponse, BaseAgent

logger = logging.getLogger(__name__)


class IntegrationAgent(BaseAgent):
    """
    Handles integration setup and FinOps optimization.

    Capabilities:
    - Integration marketplace search
    - Gateway configuration helper
    - API code generation
    """

    def __init__(self, agent_id: str = "integration_agent", transport=None):
        super().__init__(agent_id, transport)

    async def handle_message(self, message: AgentMessage) -> Optional[AgentResponse]:
        """Handle integration and FinOps tasks."""
        self.logger.info(f"IntegrationAgent handling: {message.type}")

        task_type = message.payload.get("task_type", "unknown")

        if task_type == "marketplace_search":
            return await self.search_marketplace(message.payload)
        elif task_type == "gateway_config":
            return await self.gateway_helper(message.payload)
        elif task_type == "code_generation":
            return await self.generate_integration_code(message.payload)
        else:
            return AgentResponse(
                text=f"Unknown task type: {task_type}",
                metadata={"error": True},
            )

    async def search_marketplace(self, payload: dict) -> AgentResponse:
        """
        Search integration marketplace for compatible integrations.

        BOILERPLATE: Returns mock integrations. Production should query real marketplace DB.

        Args:
            payload: Contains 'use_case' or 'keywords'

        Returns:
            AgentResponse with integration recommendations
        """
        use_case = payload.get("use_case", "")
        self.logger.info(f"Searching integrations for use case: {use_case}")

        # TODO: Implement real marketplace search
        # - Query integration directory database
        # - Filter by compatibility, pricing, ratings
        # - Rank by relevance and popularity

        mock_integrations = [
            {
                "name": "Slack Integration",
                "description": "Send notifications to Slack channels",
                "category": "communication",
                "pricing": "free",
            },
            {
                "name": "Stripe Payments",
                "description": "Process payments and subscriptions",
                "category": "payments",
                "pricing": "transaction-based",
            },
            {
                "name": "Datadog Monitoring",
                "description": "Monitor infrastructure and applications",
                "category": "monitoring",
                "pricing": "$15/host/month",
            },
        ]

        return AgentResponse(
            text=f"Found {len(mock_integrations)} integrations for '{use_case}'",
            attachments=mock_integrations,
            metadata={"use_case": use_case, "count": len(mock_integrations)},
        )

    async def gateway_helper(self, payload: dict) -> AgentResponse:
        """
        Provide API gateway configuration assistance.

        BOILERPLATE: Returns mock config. Production should generate real gateway configs.

        Args:
            payload: Contains 'integration_type', 'auth_method', etc.

        Returns:
            AgentResponse with configuration guidance
        """
        integration = payload.get("integration_type", "unknown")
        self.logger.info(f"Generating gateway config for: {integration}")

        # TODO: Generate real API gateway configuration
        # - Create route definitions
        # - Set up authentication/authorization
        # - Configure rate limiting
        # - Add monitoring/logging

        mock_config = {
            "integration": integration,
            "routes": [
                {
                    "path": f"/api/{integration}/webhook",
                    "method": "POST",
                    "auth": "api_key",
                    "rate_limit": "100/minute",
                }
            ],
            "auth_config": {
                "type": "api_key",
                "header": "X-API-Key",
            },
        }

        return AgentResponse(
            text=f"Gateway configuration generated for {integration}",
            attachments=[{"type": "gateway_config", "config": mock_config}],
            metadata={"integration": integration},
        )

    async def generate_integration_code(self, payload: dict) -> AgentResponse:
        """
        Generate boilerplate code for integrations.

        BOILERPLATE: Returns mock code. Production should use LLM code generation.

        Args:
            payload: Contains 'language', 'integration', 'framework'

        Returns:
            AgentResponse with generated code
        """
        language = payload.get("language", "python")
        integration = payload.get("integration", "unknown")

        self.logger.info(f"Generating {language} code for: {integration}")

        # TODO: Use Gemini to generate real integration code
        # - Provide integration documentation as context
        # - Generate idiomatic code in target language
        # - Include error handling and best practices

        mock_code = f"""
# Generated {language} code for {integration} integration
# BOILERPLATE: Replace with actual implementation

import requests

class {integration.title().replace('_', '')}Client:
    def __init__(self, api_key: str):
        self.api_key = api_key
        self.base_url = "https://api.{integration}.com/v1"

    def make_request(self, endpoint: str, data: dict):
        # TODO: Implement real API call
        headers = {{"Authorization": f"Bearer {{self.api_key}}"}}
        response = requests.post(f"{{self.base_url}}/{{endpoint}}", json=data, headers=headers)
        return response.json()
"""

        return AgentResponse(
            text=f"Generated {language} integration code for {integration}",
            attachments=[{"type": "code", "language": language, "code": mock_code}],
            metadata={"language": language, "integration": integration},
        )
